WIP: Refactor charts
