+++ app/src/main/java/com/ivy/wallet/ui/charts/TimeValue.kt
@@ -1,7 +1,7 @@
 package com.ivy.wallet.ui.charts
 
 import com.ivy.wallet.functional.charts.ChartPeriod
-import com.ivy.wallet.functional.charts.ChartPoint
+import com.ivy.wallet.functional.charts.SingleChartPoint
 import com.ivy.wallet.ui.onboarding.model.FromToTimeRange
 import com.ivy.wallet.ui.theme.components.charts.Value
 
@@ -11,7 +11,7 @@ data class TimeValue(
     val value: Double
 )
 
-fun List<ChartPoint>.toValues2(): List<Value> {
+fun List<SingleChartPoint>.toValues2(): List<Value> {
     return this.mapIndexed { index, it ->
         Value(
             x = index.toDouble(),
